Cleaned up includes for easier reuse
The general pattern here is to not include platform.h or config.h
from HEADERS. In other / new projects, the config.h has different
content and including it indirectly is not desirable. Do those
includes in .c files in libntech, when possible.

One complicating factor is that bool type is defined in bool.h,
which has some conditional includes. We should try to just drop
this and always include stdbool.h, but for now, a workaround is
to include stdbool.h before including whatever needs the bool
type.

Ticket: None
Changelog: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>